requirements.txt
asgiref==3.10.0
diff-match-patch==20241021
Django==5.2.7
django-filter==25.2
django-import-export==4.3.12
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
Markdown==3.9
packaging==25.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tablib==3.9.0
uritemplate==4.2.0
